Add files via upload
Problem Statement
You are the sales manager for "BeerMart", an online beer store in the United States.  You want to build a recommendation system (collaborative) for your store, where customers will be recommended the beer that they are most likely to buy. You collected the data about the ratings that the customers have provided in the past. You can download the dataset from the link below.